Update README.md
According to https://github.com/MuntashirAkon/AppManager/issues/1603#issuecomment-3208259565 , making sure not to "use ADB on both builds at the same time" doesn't suffice, testers rather need to explicitly disable the use of ADB in one build, even if going to be uninstalled, _before_ enabling it in the other. That's relevant info, it may save testers hours of pointless troubles.

Also added a sentence that helps to decide which build to use if unsure.